Add two exceptions handling registration:     - RegisterationSuccessConfirmationRequired     - RegisterationFailureUserAlreadyExisits     - Move customized login exceptions to the same package
